feat: migrate frontend to Nuxt
