+++ frontend/nuxt.config.ts
@@ -1,0 +1,7 @@
+import { defineNuxtConfig } from 'nuxt/config'
+
+export default defineNuxtConfig({
+  srcDir: 'src/',
+  css: ['~/assets/global.css', '~/assets/toast.css'],
+  components: true
+})
