+++ frontend/package.json
@@ -3,30 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "nuxt dev",
+    "build": "nuxt build",
+    "start": "nuxt start",
+    "lint": "eslint --ext .js,.vue ."
   },
   "dependencies": {
-    "core-js": "^3.8.3",
     "cropperjs": "^1.6.2",
     "echarts": "^5.6.0",
     "highlight.js": "^10.7.1",
     "ldrs": "^1.1.7",
     "markdown-it": "^14.1.0",
+    "nuxt": "^3.11.2",
     "vditor": "^3.11.1",
-    "vue": "^3.2.13",
     "vue-easy-lightbox": "^1.19.0",
     "vue-echarts": "^7.0.3",
-    "vue-router": "^4.5.1",
     "vue-toastification": "^2.0.0-rc.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
     "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
+    "@babel/preset-env": "^7.28.0",
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.0.3"
   },
