+++ frontend/src/components/CommentItem.vue
@@ -66,7 +66,7 @@ import CommentEditor from './CommentEditor.vue'
 import { renderMarkdown, handleMarkdownClick } from '../utils/markdown'
 import TimeManager from '../utils/time'
 import BaseTimeline from './BaseTimeline.vue'
-import { API_BASE_URL, toast } from '../main'
+import { API_BASE_URL, toast } from '~/main'
 import { getToken, authState } from '../utils/auth'
 import ReactionsGroup from './ReactionsGroup.vue'
 import DropdownMenu from './DropdownMenu.vue'
