+++ frontend/src/components/GlobalPopups.vue
@@ -11,7 +11,7 @@
 
 <script>
 import ActivityPopup from './ActivityPopup.vue'
-import { API_BASE_URL } from '../main'
+import { API_BASE_URL } from '~/main'
 
 export default {
   name: 'GlobalPopups',
