+++ frontend/src/pages/github-callback.vue
@@ -3,17 +3,17 @@
 </template>
 
 <script>
-import CallbackPage from '../components/CallbackPage.vue'
-import { twitterExchange } from '../utils/twitter'
+import CallbackPage from '~/components/CallbackPage.vue'
+import { githubExchange } from '~/utils/github'
 
 export default {
-  name: 'TwitterCallbackPageView',
+  name: 'GithubCallbackPageView',
   components: { CallbackPage },
   async mounted() {
     const url = new URL(window.location.href)
     const code = url.searchParams.get('code')
     const state = url.searchParams.get('state')
-    const result = await twitterExchange(code, state, '')
+    const result = await githubExchange(code, state, '')
 
     if (result.needReason) {
       this.$router.push('/signup-reason?token=' + result.token)
