+++ frontend/src/pages/message/index.vue
@@ -294,17 +294,17 @@
 <script>
 import { ref, onMounted, computed } from 'vue'
 import { useRouter } from 'vue-router'
-import { API_BASE_URL } from '../main'
-import BaseTimeline from '../components/BaseTimeline.vue'
-import BasePlaceholder from '../components/BasePlaceholder.vue'
-import NotificationContainer from '../components/NotificationContainer.vue'
-import { getToken, authState } from '../utils/auth'
-import { markNotificationsRead, fetchUnreadCount, notificationState } from '../utils/notification'
-import { toast } from '../main'
-import { stripMarkdownLength } from '../utils/markdown'
-import TimeManager from '../utils/time'
+import { API_BASE_URL } from '~/main'
+import BaseTimeline from '~/components/BaseTimeline.vue'
+import BasePlaceholder from '~/components/BasePlaceholder.vue'
+import NotificationContainer from '~/components/NotificationContainer.vue'
+import { getToken, authState } from '~/utils/auth'
+import { markNotificationsRead, fetchUnreadCount, notificationState } from '~/utils/notification'
+import { toast } from '~/main'
+import { stripMarkdownLength } from '~/utils/markdown'
+import TimeManager from '~/utils/time'
 import { hatch } from 'ldrs'
-import { reactionEmojiMap } from '../utils/reactions'
+import { reactionEmojiMap } from '~/utils/reactions'
 hatch.register()
 
 export default {
