+++ frontend/src/pages/posts/[id]/edit.vue
@@ -3,7 +3,7 @@
     <div class="new-post-form">
       <input class="post-title-input" v-model="title" placeholder="标题" />
       <div class="post-editor-container">
-        <PostEditor v-model="content" v-model:loading="isAiLoading" :disabled="!isLogin" />
+        <PostEditor v-model="content" :loading="isAiLoading" :disabled="!isLogin" />
         <LoginOverlay v-if="!isLogin" />
       </div>
       <div class="post-options">
@@ -19,17 +19,16 @@
             <i class="fa-solid fa-robot"></i>
             md格式优化
           </div>
-          <div class="post-draft" @click="saveDraft">
-            <i class="fa-solid fa-floppy-disk"></i>
-            存草稿
+          <div class="post-cancel" @click="cancelEdit">
+            取消
           </div>
           <div
             v-if="!isWaitingPosting"
             class="post-submit"
             :class="{ disabled: !isLogin }"
             @click="submitPost"
-          >发布</div>
-          <div v-else class="post-submit-loading"> <i class="fa-solid fa-spinner fa-spin"></i> 发布中...</div>
+          >更新</div>
+          <div v-else class="post-submit-loading"> <i class="fa-solid fa-spinner fa-spin"></i> 更新中...</div>
         </div>
       </div>
     </div>
@@ -38,15 +37,16 @@
 
 <script>
 import { ref, onMounted, computed } from 'vue'
-import PostEditor from '../components/PostEditor.vue'
-import CategorySelect from '../components/CategorySelect.vue'
-import TagSelect from '../components/TagSelect.vue'
-import { API_BASE_URL, toast } from '../main'
-import { getToken, authState } from '../utils/auth'
-import LoginOverlay from '../components/LoginOverlay.vue'
+import { useRoute, useRouter } from 'vue-router'
+import PostEditor from '~/components/PostEditor.vue'
+import CategorySelect from '~/components/CategorySelect.vue'
+import TagSelect from '~/components/TagSelect.vue'
+import { API_BASE_URL, toast } from '~/main'
+import { getToken, authState } from '~/utils/auth'
+import LoginOverlay from '~/components/LoginOverlay.vue'
 
 export default {
-  name: 'NewPostPageView',
+  name: 'EditPostPageView',
   components: { PostEditor, CategorySelect, TagSelect, LoginOverlay },
   setup() {
     const title = ref('')
@@ -57,82 +57,37 @@ export default {
     const isAiLoading = ref(false)
     const isLogin = computed(() => authState.loggedIn)
 
-    const loadDraft = async () => {
-      const token = getToken()
-      if (!token) return
+    const route = useRoute()
+    const router = useRouter()
+    const postId = route.params.id
+
+    const loadPost = async () => {
       try {
-        const res = await fetch(`${API_BASE_URL}/api/drafts/me`, {
-          headers: { Authorization: `Bearer ${token}` }
+        const token = getToken()
+        const res = await fetch(`${API_BASE_URL}/api/posts/${postId}`, {
+          headers: token ? { Authorization: `Bearer ${token}` } : {}
         })
-        if (res.ok && res.status !== 204) {
+        if (res.ok) {
           const data = await res.json()
           title.value = data.title || ''
           content.value = data.content || ''
-          selectedCategory.value = data.categoryId || ''
-          selectedTags.value = data.tagIds || []
-
-          toast.success('草稿已加载')
+          selectedCategory.value = data.category.id || ''
+          selectedTags.value = (data.tags || []).map(t => t.id)
         }
       } catch (e) {
-        console.error(e)
+        toast.error('加载失败')
       }
     }
 
-    onMounted(loadDraft)
+    onMounted(loadPost)
 
-    const clearPost = async () => {
+    const clearPost = () => {
       title.value = ''
       content.value = ''
       selectedCategory.value = ''
       selectedTags.value = []
-
-      // 删除草稿
-      const token = getToken()
-      if (token) {
-        const res = await fetch(`${API_BASE_URL}/api/drafts/me`, {
-          method: 'DELETE',
-          headers: {
-            Authorization: `Bearer ${token}`
-          }
-        })
-        if (res.ok) {
-          toast.success('草稿已清空')
-        } else {
-          toast.error('云端草稿清空失败, 请稍后重试')
-        }
-      }
     }
 
-    const saveDraft = async () => {
-      const token = getToken()
-      if (!token) {
-        toast.error('请先登录')
-        return
-      }
-      try {
-        const tagIds = selectedTags.value.filter(t => typeof t === 'number')
-        const res = await fetch(`${API_BASE_URL}/api/drafts`, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${token}`
-          },
-          body: JSON.stringify({
-            title: title.value,
-            content: content.value,
-            categoryId: selectedCategory.value || null,
-            tagIds
-          })
-        })
-        if (res.ok) {
-          toast.success('草稿已保存')
-        } else {
-          toast.error('保存失败')
-        }
-      } catch (e) {
-        toast.error('保存失败')
-      }
-    }
     const ensureTags = async (token) => {
       for (let i = 0; i < selectedTags.value.length; i++) {
         const t = selectedTags.value[i]
@@ -217,8 +172,8 @@ export default {
         const token = getToken()
         await ensureTags(token)
         isWaitingPosting.value = true
-        const res = await fetch(`${API_BASE_URL}/api/posts`, {
-          method: 'POST',
+        const res = await fetch(`${API_BASE_URL}/api/posts/${postId}`, {
+          method: 'PUT',
           headers: {
             'Content-Type': 'application/json',
             Authorization: `Bearer ${token}`
@@ -232,26 +187,21 @@ export default {
         })
         const data = await res.json()
         if (res.ok) {
-          if (data.reward && data.reward > 0) {
-            toast.success(`发布成功，获得 ${data.reward} 经验值`)
-          } else {
-            toast.success('发布成功')
-          }
-          if (data.id) {
-            window.location.href = `/posts/${data.id}`
-          }
-        } else if (res.status === 429) {
-          toast.error('发布过于频繁，请稍后再试')
+          toast.success('更新成功')
+          window.location.href = `/posts/${postId}`
         } else {
-          toast.error(data.error || '发布失败')
+          toast.error(data.error || '更新失败')
         }
       } catch (e) {
-        toast.error('发布失败')
+        toast.error('更新失败')
       } finally {
         isWaitingPosting.value = false
       }
     }
-    return { title, content, selectedCategory, selectedTags, submitPost, saveDraft, clearPost, isWaitingPosting, aiGenerate, isAiLoading, isLogin }
+    const cancelEdit = () => {
+      router.push(`/posts/${postId}`)
+    }
+    return { title, content, selectedCategory, selectedTags, submitPost, clearPost, cancelEdit, isWaitingPosting, aiGenerate, isAiLoading, isLogin }
   }
 }
 </script>
@@ -261,8 +211,10 @@ export default {
   display: flex;
   justify-content: center;
   background-color: var(--background-color);
+  height: 100%;
   padding-right: 20px;
   padding-left: 20px;
+  overflow-y: auto;
 }
 
 .new-post-form {
@@ -281,13 +233,13 @@ export default {
   color: var(--text-color);
 }
 
-.post-draft {
+.post-cancel {
   color: var(--primary-color);
   border-radius: 10px;
   cursor: pointer;
 }
 
-.post-draft:hover {
+.post-cancel:hover {
   text-decoration: underline;
 }
