+++ frontend/src/pages/posts/[id]/index.vue
@@ -108,21 +108,21 @@
 import { ref, computed, onMounted, onBeforeUnmount, nextTick, watch } from 'vue'
 import VueEasyLightbox from 'vue-easy-lightbox'
 import { useRoute } from 'vue-router'
-import CommentItem from '../components/CommentItem.vue'
-import CommentEditor from '../components/CommentEditor.vue'
-import BaseTimeline from '../components/BaseTimeline.vue'
-import ArticleTags from '../components/ArticleTags.vue'
-import ArticleCategory from '../components/ArticleCategory.vue'
-import ReactionsGroup from '../components/ReactionsGroup.vue'
-import DropdownMenu from '../components/DropdownMenu.vue'
-import { renderMarkdown, handleMarkdownClick, stripMarkdownLength } from '../utils/markdown'
-import { API_BASE_URL, toast } from '../main'
-import { getToken, authState } from '../utils/auth'
-import TimeManager from '../utils/time'
+import CommentItem from '~/components/CommentItem.vue'
+import CommentEditor from '~/components/CommentEditor.vue'
+import BaseTimeline from '~/components/BaseTimeline.vue'
+import ArticleTags from '~/components/ArticleTags.vue'
+import ArticleCategory from '~/components/ArticleCategory.vue'
+import ReactionsGroup from '~/components/ReactionsGroup.vue'
+import DropdownMenu from '~/components/DropdownMenu.vue'
+import { renderMarkdown, handleMarkdownClick, stripMarkdownLength } from '~/utils/markdown'
+import { API_BASE_URL, toast } from '~/main'
+import { getToken, authState } from '~/utils/auth'
+import TimeManager from '~/utils/time'
 import { hatch } from 'ldrs'
 import { useRouter } from 'vue-router'
-import { isMobile } from '../utils/screen'
-import Dropdown from '../components/Dropdown.vue'
+import { isMobile } from '~/utils/screen'
+import Dropdown from '~/components/Dropdown.vue'
 hatch.register()
 
 export default {
