+++ frontend/src/pages/signup-reason.vue
@@ -17,8 +17,8 @@
 </template>
 
 <script>
-import BaseInput from '../components/BaseInput.vue'
-import {API_BASE_URL, toast} from '../main'
+import BaseInput from '~/components/BaseInput.vue'
+import {API_BASE_URL, toast} from '~/main'
 
 export default {
   name: 'SignupReasonPageView',
