+++ frontend/src/pages/twitter-callback.vue
@@ -3,17 +3,17 @@
 </template>
 
 <script>
-import CallbackPage from '../components/CallbackPage.vue'
-import { discordExchange } from '../utils/discord'
+import CallbackPage from '~/components/CallbackPage.vue'
+import { twitterExchange } from '~/utils/twitter'
 
 export default {
-  name: 'DiscordCallbackPageView',
+  name: 'TwitterCallbackPageView',
   components: { CallbackPage },
   async mounted() {
     const url = new URL(window.location.href)
     const code = url.searchParams.get('code')
     const state = url.searchParams.get('state')
-    const result = await discordExchange(code, state, '')
+    const result = await twitterExchange(code, state, '')
 
     if (result.needReason) {
       this.$router.push('/signup-reason?token=' + result.token)
