+++ frontend/src/plugins/directives.client.js
@@ -1,0 +1,6 @@
+import { defineNuxtPlugin } from '#app'
+import ClickOutside from '~/directives/clickOutside'
+
+export default defineNuxtPlugin((nuxtApp) => {
+  nuxtApp.vueApp.directive('click-outside', ClickOutside)
+})
