--- frontend/vue.config.js
@@ -1,11 +1,0 @@
-const { defineConfig } = require('@vue/cli-service')
-module.exports = defineConfig({
-    // publicPath: process.env.NODE_ENV === 'production'
-    // // ? '/ai-cv/' 
-    // ? '/OpenIsle/' 
-    // : '/',
-  transpileDependencies: true,
-  configureWebpack: {
-    devtool: 'source-map'
-  },
-})
